+++ optimizations/JPEG/Images/Pixel.cs
@@ -7,7 +7,7 @@ public class Pixel
 {
 	private readonly PixelFormat format;
 
-	public Pixel(double firstComponent, double secondComponent, double thirdComponent, PixelFormat pixelFormat)
+	public Pixel(float firstComponent, float secondComponent, float thirdComponent, PixelFormat pixelFormat)
 	{
 		switch (pixelFormat)
 		{
@@ -27,16 +27,16 @@ public class Pixel
 		format = pixelFormat;
 	}
 
-	private readonly double r, g, b, y, cb, cr;
+	private readonly float r, g, b, y, cb, cr;
 
-	public double R => format == PixelFormat.RGB ? r : (298.082 * y + 408.583 * Cr) / 256.0 - 222.921;
+	public float R => format == PixelFormat.RGB ? r : (298.082f * y + 408.583f * Cr) / 256.0f - 222.921f;
 
-	public double G =>
-		format == PixelFormat.RGB ? g : (298.082 * Y - 100.291 * Cb - 208.120 * Cr) / 256.0 + 135.576;
+	public float G =>
+		format == PixelFormat.RGB ? g : (298.082f * Y - 100.291f * Cb - 208.120f * Cr) / 256.0f + 135.576f;
 
-	public double B => format == PixelFormat.RGB ? b : (298.082 * Y + 516.412 * Cb) / 256.0 - 276.836;
+	public float B => format == PixelFormat.RGB ? b : (298.082f * Y + 516.412f * Cb) / 256.0f - 276.836f;
 
-	public double Y => format == PixelFormat.YCbCr ? y : 16.0 + (65.738 * R + 129.057 * G + 24.064 * B) / 256.0;
-	public double Cb => format == PixelFormat.YCbCr ? cb : 128.0 + (-37.945 * R - 74.494 * G + 112.439 * B) / 256.0;
-	public double Cr => format == PixelFormat.YCbCr ? cr : 128.0 + (112.439 * R - 94.154 * G - 18.285 * B) / 256.0;
+	public float Y => format == PixelFormat.YCbCr ? y : 16.0f + (65.738f * R + 129.057f * G + 24.064f * B) / 256.0f;
+	public float Cb => format == PixelFormat.YCbCr ? cb : 128.0f + (-37.945f * R - 74.494f * G + 112.439f * B) / 256.0f;
+	public float Cr => format == PixelFormat.YCbCr ? cr : 128.0f + (112.439f * R - 94.154f * G - 18.285f * B) / 256.0f;
 }
